mysite/requirements.txt
asgiref
bleach
certifi
charset-normalizer
django-appconf
django-compressor
django-debug-toolbar
django-extensions
django-multi-captcha-admin
django-ranged-response
django-recaptcha2
django-robots
django-simple-captcha
django-summernote
django-taggit
idna
pillow
rcssmin
requests
rjsmin
sqlparse
tzdata
urllib3
webencodings
